- Standarize date style operations with constants - Move away any fuction definition into a module sqlquery - Enhance MatchHistoryETL readability -
