+++ ETL/Python ETL/MatchHistoryETL.py
@@ -14,82 +14,14 @@
 # os.getcwd()
 
 import constants
-import requests
+import sqlquery
 import pandas as pd
 import numpy as np
 import psycopg2
 import csv
-import datetime
-from psycopg2 import sql
-import csv
 
 pd.set_option('display.max_rows', 100)
 pd.set_option('display.max_columns', 50)
-
-
-# In[ ]:
-
-
-def calculate_matchweek(conn):
-    """
-        Finds out last registered matchweek for current season in Database. 
-        Returns last registerered matchweek plus one. If there's any problem 
-        with databse or season has not even started, then returns default matchweek "1"
-    """
-    try:
-        cursor = conn.cursor()
-        matchweek_query = f"SELECT MAX(\"MatchWeek\") FROM public.match_history WHERE \"Season\" = '{constants.CURRENT_SEASON_TAG}'"
-        cursor.execute(matchweek_query)
-        matchweek_result = cursor.fetchone()
-        if matchweek_result[0] is None:
-            return constants.DEFAULT_MATCHWEEK
-        else:
-            # Next matchweek
-            return matchweek_result[0] + 1
-    except psycopg2.Error as e:
-        print ("An error ocurred in database")
-        print (e)
-        return constants.DEFAULT_MATCHWEEK
-    finally:
-        cursor.close()
-
-
-# In[ ]:
-
-
-def upsert_query(to_table, reference_column, dataframe):
-    """
-        Generates a Composed SQL object that contains all necessary SQL code to 
-        perform upserts into a postgresql table from dataframe as source.
-        If row attempted to insert is already present in database, then engine
-        will skip insertion. 
-
-        Input:
-            to_table - target table name in format 'schema.table_name' to generate inserts 
-            reference_column - Column name of your table that is considered PK to avoid overwritings
-            dataframe - pandas dataframe containing all rows to be inserted
-    """
-    date_config = sql.SQL('SET datestyle = "ISO, DMY"; ')
-    
-    insert_header = sql.SQL("INSERT INTO " + to_table + " ({fields})").format(
-        fields = sql.SQL(',').join([sql.Identifier(column) for column in dataframe.columns])
-    )
-    
-    values_array = []
-    for row in dataframe.iterrows():
-        values_array.append(sql.SQL(" ({values})").format(
-            values = sql.SQL(',').join([sql.Literal(value) for value in row[1].values])
-        ))
-    
-    insert_values = sql.SQL(' VALUES ') + sql.SQL(',').join(values_array)
-    
-    insert_condition = sql.SQL(" ON CONFLICT({conflict_col}) DO NOTHING").format(
-        conflict_col = sql.Identifier(reference_column)
-    )
-    
-    insert_statement = date_config + insert_header + insert_values + insert_condition
-   
-    return insert_statement
 
 
 # In[ ]:
@@ -197,7 +129,7 @@ else:
 
 
 # Find out current season matchweek
-next_matchweek = calculate_matchweek(connection)
+next_matchweek = sqlquery.calculate_matchweek(connection)
 
 
 # In[ ]:
@@ -233,7 +165,7 @@ website_frame["AwayTeamPoints"] = np.select(conditions, away_points)
 
 # Keep up to date postgresql database
 
-insert_statement = upsert_query("public.match_history", "MatchID", website_frame)
+insert_statement = sqlquery.upsert_query(to_table="public.match_history", reference_column="MatchID",dataframe=website_frame)
 
 
 # In[ ]:
@@ -258,22 +190,17 @@ finally:
 
 # Keep up to date master dataset
 try:
-    with connection.cursor() as cursor:
-        date_config = sql.SQL('SET datestyle = "ISO, DMY"; ')
-        select_statement = sql.SQL("SELECT * FROM public.match_history")
-        final_statement = date_config + select_statement
-        cursor.execute(final_statement)
-        results = cursor.fetchall()
-        with(open(constants.MASTER_CSV_URL, mode='w',newline="") as file):
-            writer = csv.writer(file)
-            # write column names
-            writer.writerow([column_description[0] for column_description in cursor.description])
-            writer.writerows(results)
+    rowcount, columns, rows = sqlquery.select_all_matches(connection)
+    with(open(constants.MASTER_CSV_URL, mode='w',newline="") as file):
+        writer = csv.writer(file)
+        # write column names
+        writer.writerow(columns)
+        writer.writerows(rows)
 except psycopg2.Error as e:
     print (f'Can not connect to the postgress database "{constants.DB_NAME}". Make sure database server is running')
     print (e)
 else:
-    print (f"Rows in Master CSV Dataset: {cursor.rowcount}")
+    print (f"Rows in Master CSV Dataset: {rowcount}")
 finally:
     cursor.close()
             
@@ -283,4 +210,16 @@ finally:
 
 
 # Keep up to date Kaggle dataset
+
+
+# In[ ]:
+
+
+connection.close()
+
+
+# In[ ]:
+
+
+
