+++ ETL/Python ETL/constants.py
@@ -13,3 +13,4 @@ DB_PORT = "5432"
 DB_USER = "admin"
 DB_PASSWORD = "root"
 DB_NAME = "premier_league"
+DB_DATE_STYLE = "ISO, DMY"
